pages for inbox with redirect and today folders and sidenav for components folder Co-authored-by: Luis Carlin <luiscarlin@gmail.com>
