+++ app/layout.js
@@ -1,4 +1,5 @@
 import { Inter } from "next/font/google";
+import Sidebar from "./components/sidebar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -10,7 +11,10 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Sidebar />
+        <main>{children}</main>
+      </body>
     </html>
   );
 }
