+++ app/today/page.js
@@ -1,0 +1,18 @@
+export default function TodayPage() {
+  return <div>this is the today page</div>;
+}
+
+/*
+
+
+app/
+    page.js
+    layout.js
+    inbox/
+        page.js
+        layout.js
+
+
+
+
+*/
